Struct mishap::Mishap

pub struct Mishap { /* private fields */ }

A generic tree of errors, where each error can have any number of sources.

Like anyhow::Error, but with a tree structure.

For example:

An anyhow::Error or other Rust error is logically a chain of errors.

A Vec<anyhow::Error> is a list of errors, each of which is a chain.

Implementations

impl Mishap

pub fn from_msg<D>(msg: D) -> Self
where D: Debug + Display + Send + Sync + 'static,

pub fn from_anyhow(error: Error) -> Self

pub fn from_msg_and_anyhow<D>(msg: D, error: Error) -> Self
where D: Display + Send + Sync + 'static,

pub fn from_msg_and_anyhows<D, I>(msg: D, sources: I) -> Self
where D: Display + Send + Sync + 'static, I: IntoIterator<Item = Error>,

pub fn from_error<E>(error: E) -> Self
where E: Error + Send + Sync + 'static,

pub fn from_msg_and_error<D, E>(msg: D, error: E) -> Self
where D: Display + Send + Sync + 'static, E: Error + Send + Sync + 'static,

pub fn from_msg_and_errors<D, I, E>(msg: D, sources: I) -> Self
where D: Display + Send + Sync + 'static, I: IntoIterator<Item = E>, E: Error + Send + Sync + 'static,

pub fn from_error_tree<ET>(tree: ET) -> Self
where ET: ErrorTree + 'static,

pub fn from_msg_and_error_tree<D, ET>(msg: D, tree: ET) -> Self
where D: Display + Send + Sync + 'static, ET: ErrorTree + 'static,

pub fn from_msg_and_error_trees<D, I, ET>(msg: D, sources: I) -> Self
where D: Display + Send + Sync + 'static, I: IntoIterator<Item = ET>, ET: ErrorTree + 'static,

pub fn from_borrowed_error(error: &dyn Error) -> Self

Constructs a tree from a borrowed error, effectively cloning it by stringifying it.

This doesn’t currently preserve Debug information.

pub fn from_borrowed_tree(tree: &dyn ErrorTree) -> Self

Constructs a tree from a borrowed tree, effectively cloning it by stringifying it.

This doesn’t currently preserve Debug information.

pub fn from_msg_and_cause_chain<I, D>(msg: D, cause_chain: I) -> Self
where I: DoubleEndedIterator<Item = D>, D: Display + Debug + Send + Sync + 'static,

Create a tree from a message and a chain of causes.

The iterator represents a chain of causes rather than siblings.

pub fn wrap_mishap<D>(self, msg: D) -> Self
where D: Display + Send + Sync + 'static,